requirements.txt
openai
google-generativeai
crewai
python-dotenv
pyautogui
python-pptx
playwright
requests
google-genai
requests
lxml